requirements.txt
aiogram==3.4.1
fastapi==0.115.2
uvicorn==0.30.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
python-dotenv==1.0.1
